requirements.txt
torch  --index-url https://download.pytorch.org/whl/cu121
pandas
matplotlib
joblib
tqdm
transformers
datasets
torchmetrics